pythonapp/requirements.txt
fastapi==0.104.1
python-socketio==5.10.0
uvicorn==0.24.0
aiohttp==3.8.6
